feat: 양의 실수도 가능하게 로직 변경

+++ src/main/java/calculator/domain/CalculatorModel.java
@@ -7,7 +7,7 @@ public class CalculatorModel {
 
   private static final String DEFAULT_DELIMITER = ",|:";
 
-  public int sum(String input) {
+  public double sum(String input) {
     if (input.isBlank()) {
       return 0;
     }
